RandomRange tests in a separate class.
